Use BUFGs for some reset signals
Free up some data-routing resource by using global clock-routing
resources for some specific high-level reset signals.
Not perfect as it drops down to normal data-routing when
it gets inverted for use in many places, but does seem to
improve timing and routing utilisation a bit.
Might as well, seeing as we were only using 8 of 24 BUFGs.